fix: context-aware completion, semantic-identity tokens, and UTF-16 positions
Completion now uses the requested position: member context offers enum
members or corpus-evidenced receiver methods, other contexts filter
declared symbols/builtins/keywords by the typed prefix. Semantic tokens
classify by the semantic index spans and symbol kinds (not name-string
membership), and document/span position conversion is UTF-16-aware with
non-BMP regression tests.

+++ crates/wright-lsp/src/main.rs
@@ -373,10 +373,7 @@ fn publish_diagnostics(
     let params = PublishDiagnosticsParams {
         uri: url,
         diagnostics,
-        version: service
-            .store
-            .document(uri)
-            .map(|document| document.version as i32),
+        version: service.store.document(uri).map(|document| document.version),
     };
     let notification = serde_json::json!({
         "jsonrpc": "2.0",
